courses: change credits to float
* Allow for half credit courses

+++ db/migrate/20160407165307_create_courses.rb
@@ -3,7 +3,7 @@ class CreateCourses < ActiveRecord::Migration
     create_table :courses do |t|
       t.text :title
       t.text :instructor
-      t.integer :credits
+      t.float :credits
       t.text :description
       t.text :requirements
       t.text :term
